Chore: 계층 구조 변경
- 엔터티 파일명 "Entity" 명시
- DTO 내 request, response 패키징 처리
