--- src/main/java/fish/common/bungbal/dto/BungbalRequest.java
@@ -1,8 +1,0 @@
-package fish.common.bungbal.dto;
-
-import lombok.Getter;
-
-@Getter
-public class BungbalRequest {
-    private String mbti;
-}
