+++ src/main/java/fish/common/bungbal/dto/request/BungbalRequest.java
@@ -1,0 +1,8 @@
+package fish.common.bungbal.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class BungbalRequest {
+    private String mbti;
+}
