+++ src/main/java/fish/common/calendar/dto/request/CalendarModifyRequest.java
@@ -1,0 +1,9 @@
+package fish.common.calendar.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class CalendarModifyRequest {
+    private Long calendarId;
+    private String contents;
+}
