--- src/main/java/fish/common/calendar/request/CalendarModifyRequest.java
@@ -1,9 +1,0 @@
-package fish.common.calendar.request;
-
-import lombok.Getter;
-
-@Getter
-public class CalendarModifyRequest {
-    private Long calendarId;
-    private String contents;
-}
