+++ src/main/java/fish/common/community/post/repository/PostRepository.java
@@ -1,9 +1,9 @@
 package fish.common.community.post.repository;
 
-import fish.common.community.post.entity.Post;
+import fish.common.community.post.entity.PostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
 }
