--- src/main/java/fish/common/detail/dto/DetailRequest.java
@@ -1,32 +1,0 @@
-package fish.common.detail.dto;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import fish.common.detail.entity.DetailEntity;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.Arrays;
-import java.util.List;
-
-@Data
-@RequiredArgsConstructor
-public class DetailRequest {
-    private String flavors;
-    private MultipartFile picture;
-    private String date;
-    private String contents;
-
-    ObjectMapper objectMapper = new ObjectMapper();
-
-    public DetailEntity toEntity(DetailRequest request, Long userId) throws JsonProcessingException {
-        List<DetailFlavor> detailFlavors = Arrays.asList(objectMapper.readValue(request.flavors, DetailFlavor[].class));
-        return DetailEntity.builder()
-                .flavors(detailFlavors)
-                .userId(userId)
-                .date(date)
-                .contents(contents)
-                .build();
-    }
-}
