+++ src/main/java/fish/common/flavor/dto/Status.java
@@ -1,0 +1,5 @@
+package fish.common.flavor.dto;
+
+public enum Status {
+    APPROVED, PENDING, DENIED
+}
