+++ src/main/java/fish/common/flavor/dto/request/FlavorReportRequest.java
@@ -1,0 +1,10 @@
+package fish.common.flavor.dto.request;
+
+import lombok.Getter;
+import jakarta.validation.constraints.NotNull;
+
+@Getter
+public class FlavorReportRequest {
+    @NotNull
+    private String flavors;
+}
