--- src/main/java/fish/common/flavor/entity/FishBunFlavor.java
@@ -1,27 +1,0 @@
-package fish.common.flavor.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Table(name = "FISH_BUN_FLAVOR")
-@Getter
-@Entity
-@NoArgsConstructor
-public class FishBunFlavor {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String flavor;
-    private String iconCode;
-    private int seq;
-    @CreationTimestamp
-    private LocalDateTime regDate;
-    private String description;
-    private String highlight;
-}
