--- src/main/java/fish/common/flavor/entity/Status.java
@@ -1,5 +1,0 @@
-package fish.common.flavor.entity;
-
-public enum Status {
-    APPROVED, PENDING, DENIED
-}
