+++ src/main/java/fish/common/flavor/repository/FlavorReportRepository.java
@@ -1,6 +1,6 @@
 package fish.common.flavor.repository;
 
-import fish.common.flavor.entity.FishBunFlavorReport;
+import fish.common.flavor.entity.FishBunFlavorReportEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -9,7 +9,7 @@ import java.util.Map;
 import java.util.List;
 
 @Repository
-public interface FlavorReportRepository extends JpaRepository<FishBunFlavorReport, Long> {
+public interface FlavorReportRepository extends JpaRepository<FishBunFlavorReportEntity, Long> {
     @Query(value = "SELECT FBFR.flavor AS flavor, FBFR.status AS status, U.nickname AS nickname " +
             "FROM FISH_BUN_FLAVOR_REPORT FBFR " +
             "LEFT JOIN USER U ON FBFR.userId = U.id", nativeQuery = true)
