+++ src/main/java/fish/common/flavor/repository/FlavorRepository.java
@@ -1,14 +1,14 @@
 package fish.common.flavor.repository;
 
-import fish.common.flavor.entity.FishBunFlavor;
+import fish.common.flavor.entity.FishBunFlavorEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface FlavorRepository extends JpaRepository<FishBunFlavor, Long> {
-    List<FishBunFlavor> findAllByOrderBySeqAsc();
+public interface FlavorRepository extends JpaRepository<FishBunFlavorEntity, Long> {
+    List<FishBunFlavorEntity> findAllByOrderBySeqAsc();
 
-    FishBunFlavor findIconCodeById(Long id);
+    FishBunFlavorEntity findIconCodeById(Long id);
 }
