--- src/main/java/fish/common/flavor/request/FlavorReportRequest.java
@@ -1,10 +1,0 @@
-package fish.common.flavor.request;
-
-import lombok.Getter;
-import jakarta.validation.constraints.NotNull;
-
-@Getter
-public class FlavorReportRequest {
-    @NotNull
-    private String flavors;
-}
