+++ src/main/java/fish/common/main/dto/response/FishBunDayCountResponse.java
@@ -1,0 +1,13 @@
+package fish.common.main.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import java.util.Map;
+
+@Getter
+@AllArgsConstructor
+public class FishBunDayCountResponse {
+    private Map<String, Object> daysInWeek;
+    private int weeklyCount;
+    private int monthlyCount;
+}
