--- src/main/java/fish/common/user/dto/UserRequest.java
@@ -1,10 +1,0 @@
-package fish.common.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserRequest {
-    private String nickname;
-}
