+++ src/main/java/fish/common/user/dto/request/UserRequest.java
@@ -1,0 +1,10 @@
+package fish.common.user.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserRequest {
+    private String nickname;
+}
